fix: split server-only jiti evaluation into separate 'use server' file
Move evaluateConfigContent() and Node.js imports (fs, os, path, jiti) to
github-config-push-server.ts marked with 'use server' so Next.js keeps
them out of the client bundle. Fixes Turbopack 'Can't resolve fs' error.
